Delete user.

+++ internal/trasport/http/handler/userHandler.go
@@ -3,6 +3,7 @@ package handler
 import (
 	"net/http"
 	"strconv"
+	"vado_server/internal/config/code"
 	"vado_server/internal/domain/user"
 
 	"github.com/gin-gonic/gin"
@@ -38,12 +39,42 @@ func DeleteUser(service *user.Service) func(c *gin.Context) {
 			return
 		}
 
+		currentId, errGetUSerId := c.Get(code.UserId)
+		if !errGetUSerId {
+			c.JSON(500, gin.H{
+				"error": "Not fount user id",
+			})
+			return
+		}
+
+		if currentId == uint(parseUint) {
+			c.JSON(400, gin.H{
+				"error": "Вы пытаетесь удалить свой собственный аккаунт",
+			})
+			return
+		}
+
+		role, errRole := c.Get(code.Role)
+		if !errRole {
+			c.JSON(500, gin.H{
+				"error": "Not fount user role",
+			})
+			return
+		}
+
+		if role != "admin" {
+			c.JSON(400, gin.H{
+				"error": "У вас нет прав на удаление аккаунтов",
+			})
+			return
+		}
+
 		err := service.DeleteUser(uint(parseUint))
 		if err != nil {
 			ShowError(c, "Failed to delete user", err.Error())
 			return
 		}
 
-		c.JSON(200, gin.H{"status": "deleted"})
+		c.JSON(200, gin.H{})
 	}
 }
